Moving Character\Domain\Entities\ValueObjects to Character\Domain\ValueObjects
