+++ app/Modules/Character/Domain/Factories/CharacterFactory.php
@@ -3,15 +3,15 @@
 
 namespace App\Modules\Character\Domain\Factories;
 
-use App\Modules\Character\Domain\Entities\ValueObjects\HitPoints;
 use App\Traits\GeneratesUuid;
 use App\Modules\Character\Domain\Contracts\RaceRepositoryInterface;
 use App\Modules\Character\Domain\Entities\Attributes;
 use App\Modules\Character\Domain\Entities\Character;
 use App\Modules\Character\Domain\Entities\Gender;
-use App\Modules\Character\Domain\Entities\Money;
-use App\Modules\Character\Domain\Entities\ValueObjects\Reputation;
-use App\Modules\Character\Domain\Entities\ValueObjects\Xp;
+use App\Modules\Character\Domain\ValueObjects\HitPoints;
+use App\Modules\Character\Domain\ValueObjects\Money;
+use App\Modules\Character\Domain\ValueObjects\Reputation;
+use App\Modules\Character\Domain\ValueObjects\Xp;
 use App\Modules\Character\Domain\Requests\CreateCharacterRequest;
