+++ app/Modules/Character/Domain/ValueObjects/HitPoints.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace App\Modules\Character\Domain\Entities\ValueObjects;
+namespace App\Modules\Character\Domain\ValueObjects;
 
 
 use App\Modules\Character\Domain\Entities\Race;
