+++ app/Modules/Character/Domain/ValueObjects/Money.php
@@ -1,10 +1,10 @@
 <?php
 
 
-namespace App\Modules\Character\Domain\Entities\ValueObjects;
+namespace App\Modules\Character\Domain\ValueObjects;
 
 
-class Reputation
+class Money
 {
     /**
      * @var int
