+++ app/Modules/Character/Domain/ValueObjects/Name.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace App\Modules\Character\Domain\Entities\ValueObjects;
+namespace App\Modules\Character\Domain\ValueObjects;
 
 
 class Name
